+++ get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: lbourniq <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 12:08:49 by lbourniq          #+#    #+#             */
-/*   Updated: 2022/11/19 13:40:19 by lbourniq         ###   ########lyon.fr   */
+/*   Updated: 2022/11/19 13:52:03 by lbourniq         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!str)
